\docType{data}
\name{ESI}
\alias{ESI}
\title{Employee Satisfaction Index (ESI) Data Set}
\format{
A data frame with 500 rows and 10 features. 
\describe{
 \item{emp_id}{label.}
  \item{age}{continuous from 23 to 45.}
  \item{Dept}{categorical.}
  \item{location}{binary.}
  \item{education}{binary.}
  \item{recruitment_type}{categorical.}
  \item{job_level}{ordinal from 1 to 5.}
 \item{rating}{ordinal from 1 to 5.}
  \item{onsite}{binary.}
  \item{awards}{number of awards 0-9.}
  \item{certifications}{binary.}
  \item{salary}{continuous from 24.1  to 86.8.}
  \item{satisfied}{binary.}
  }}
\source{
Harris, M. (2023). Employee Satisfaction Index Dataset. Evanston, Illinois: Kaggle. Version 1
}
\usage{
data(ESI)
}
\description{ The Employee Satisfaction Index (ESI) data set, from Kaggle (Harris, 2023), is a fictional data set
that measures employee satisfaction}
\keyword{datasets}
